requirements.txt
streamlit~=1.40
langgraph~=0.2
langchain~=0.3
langchain-openai~=0.2
langchain-anthropic~=0.3
azure-ai-documentintelligence~=1.0
azure-identity~=1.19
pydantic~=2.9
pydantic-settings~=2.6
python-dotenv~=1.0
graphviz>=0.20
